+++ source/selery/about.d
@@ -127,10 +127,6 @@ const struct Software {
 	enum string display = name ~ "/" ~ displayVersion;
 
 	enum string simpleDisplay = name ~ " " ~ to!string(major) ~ "." ~ to!string(minor) ~ (patch != 0 ? "." ~ to!string(patch) : "");
-
-	// HNCOM
-
-	enum uint hncomProtocol = 10;
 	
 	/+/**
 	 * Version of the api used by the software. It's used to check the